refactor: move scripts as commands to a standalone dir (#18)
Some scripts that are meant to be executed as commands, they were
previously kept indiscriminately with other scripts under "./scripts/".
Now they are moved to "./scripts/as-commands/" directory.

The directory "/etc/nixos/scripts/as-commands/" (which is a symlink of
"./scripts/as-commands/") is added to the PATH environment variable in
each shell, including zsh, bash, nu.

This change eliminates the possibility to inadvertently execute scripts
that are not intended to be used as a globally available command.